refactor(js_interop_gen): extract pure utilities and data classes
Cleans up transformer.dart by moving isolated logic and pure utility
functions that do not rely on the core transformer state.

- Extracted `ExportReference` data class into a dedicated `export_reference.dart` file.
- Extracted context-independent helper functions (`parseModifiers`, `parseQualifiedNameParts`, `parseQualifiedNameFromTSQualifiedName`, `parseQualifiedName`, and `toCamelCase`) into a new `utils.dart` file.
- Inlined the short `parseNumericLiteral` and `parseStringLiteral` helpers directly at their call sites in the `Transformer` class.
- Updated imports, made the extracted utility functions public, and cleaned up unused imports.

No functional changes: All tests pass, and package:web code generation yields a zero diff.
